Fix linter errors and improve project structure
- Fixed import errors by replacing greatestworks/aop/logger with greatestworks/internal/infrastructure/logger
- Fixed syntax errors including missing closing braces and parentheses
- Fixed redeclared errors by commenting out duplicate struct definitions in protocol files
- Fixed undefined types and variables by adding missing imports and error definitions
- Fixed method signature mismatches in player service and repository implementations
- Removed gRPC dependencies and integrated netcore-go RPC framework
- Updated spoor logging to v2.0.1 with singleton pattern
- Fixed protocol message redeclarations by consolidating definitions
- Improved error handling and type safety across the codebase
- Reduced build errors from 2443+ to manageable set of remaining issues

+++ internal/interfaces/tcp/protocol/game_protocol.go
@@ -436,22 +436,22 @@ func IsValidMessageType(msgType uint32) bool {
 // GetMessageTypeName 获取消息类型名称
 func GetMessageTypeName(msgType uint32) string {
 	msgNames := map[uint32]string{
-		MsgPlayerLogin:      "PlayerLogin",
-		MsgPlayerLogout:     "PlayerLogout",
-		MsgPlayerMove:       "PlayerMove",
-		MsgPlayerInfo:       "PlayerInfo",
-		MsgPlayerCreate:     "PlayerCreate",
-		MsgCreateBattle:     "CreateBattle",
-		MsgJoinBattle:       "JoinBattle",
-		MsgStartBattle:      "StartBattle",
-		MsgBattleAction:     "BattleAction",
-		MsgLeaveBattle:      "LeaveBattle",
-		MsgGetPlayerInfo:    "GetPlayerInfo",
-		MsgGetOnlinePlayers: "GetOnlinePlayers",
-		MsgGetBattleInfo:    "GetBattleInfo",
-		MsgHeartbeat:        "Heartbeat",
-		MsgPing:             "Ping",
-		MsgError:            "Error",
+		MsgPlayerLogin:       "PlayerLogin",
+		MsgPlayerLogout:      "PlayerLogout",
+		MsgPlayerMove:        "PlayerMove",
+		MsgPlayerInfo:        "PlayerInfo",
+		MsgPlayerCreate:      "PlayerCreate",
+		MsgCreateBattle:      "CreateBattle",
+		MsgJoinBattle:        "JoinBattle",
+		MsgStartBattle:       "StartBattle",
+		MsgBattleAction:      "BattleAction",
+		MsgLeaveBattle:       "LeaveBattle",
+		MsgGetPlayerInfo:     "GetPlayerInfo",
+		MsgGetOnlinePlayers:  "GetOnlinePlayers",
+		MsgGetBattleInfo:     "GetBattleInfo",
+		uint32(MsgHeartbeat): "Heartbeat",
+		uint32(MsgPing):      "Ping",
+		uint32(MsgError):     "Error",
 	}
 
 	if name, exists := msgNames[msgType]; exists {
